Flash Drive Pictures folder
Run file 7. Press RETURN key to show all pictures.
With your latest Flash Drive code, my 4052 ran this fine.
My 4054A hung on the READ in line 620 on random pictures.  RUN 620 did not read any additional info.  Flash drive was hung - I power cycled and tried again multiple times

Your latest 405x emulator runs the program - but the INPUT @5,19: to get the file header is not working right.  It does not get the file header from the prior FIND, it appears to just increment the file that it is using to get the header.